requirements.txt
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.2
fastapi==0.109.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.0
pytest-asyncio==0.23.5
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.27.0
# uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
